+++ packages/deepsea-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # deepsea-components
 
+## 5.17.10
+
+### Patch Changes
+
+- 新增 useEventListener，升级 react 最低版本至 19.2.0
+- Updated dependencies
+    - soda-hooks@6.16.0
+
 ## 5.17.9
 
 ### Patch Changes