+++ packages/react-soda/CHANGELOG.md
@@ -1,5 +1,11 @@
 # react-soda
 
+## 5.3.2
+
+### Patch Changes
+
+- 新增 useEventListener，升级 react 最低版本至 19.2.0
+
 ## 5.3.1
 
 ### Patch Changes