+++ packages/soda-antd/CHANGELOG.md
@@ -1,5 +1,14 @@
 # soda-antd
 
+## 0.12.5
+
+### Patch Changes
+
+- 新增 useEventListener，升级 react 最低版本至 19.2.0
+- Updated dependencies
+    - soda-hooks@6.16.0
+    - soda-react@0.1.5
+
 ## 0.12.4
 
 ### Patch Changes